make everything but extension api internal
