+++ dotnet/src/Experimental/Experimental.Prompty.UnitTests/TestData/prompties/chat.prompty
