+++ dotnet/src/Experimental/Experimental.Prompty.UnitTests/TestData/prompties/coherence.prompty
