+++ dotnet/src/Experimental/Experimental.Prompty.UnitTests/TestData/prompties/groundedness.prompty
