+++ dotnet/src/Extensions/PromptTemplates.Prompty/Core/Types/ApiType.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+namespace Microsoft.SemanticKernel.Experimental.Prompty.Core;
+
+internal enum ApiType
+{
+    Chat,
+    Completion,
+    Image,
+    Embedding
+}
