+++ dotnet/src/Extensions/PromptTemplates.Prompty/Core/Types/ModelType.cs
@@ -1,0 +1,8 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+namespace Microsoft.SemanticKernel.Experimental.Prompty.Core;
+
+internal enum ModelType
+{
+    azure
+}
