+++ dotnet/src/Extensions/PromptTemplates.Prompty/Core/Types/ParserType.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+namespace Microsoft.SemanticKernel.Experimental.Prompty.Prompty;
+
+internal enum ParserType
+{
+    Chat,
+    Embedding,
+    Completion,
+    Image,
+}
