+++ dotnet/src/Extensions/PromptTemplates.Prompty/Core/Types/RoleType.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+namespace Microsoft.SemanticKernel.Experimental.Prompty.Core;
+
+internal enum RoleType
+{
+    assistant,
+    function,
+    system,
+    tool,
+    user,
+}
