+++ dotnet/src/Extensions/PromptTemplates.Prompty/Extensions/PromptyKernelExtension.cs
@@ -1,14 +1,15 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
-using Prompty.Core.Parsers;
-using Prompty.Core.Renderers;
-namespace Microsoft.SemanticKernel.PromptTemplates.Prompty.Extensions;
+namespace Microsoft.SemanticKernel.Experimental.Prompty.Extension;
+
 public static class PromptyKernelExtension
 {
     public static KernelFunction CreateFunctionFromPrompty(
         this Kernel _,
-        global::Prompty.Core.Prompty prompty)
+        string promptyPath)
     {
+        var prompty = new Core.Prompty();
+        prompty = prompty.Load(promptyPath, prompty);
         var promptFunction = new PromptyKernelFunction(prompty);
 
         return promptFunction;
