--- dotnet/src/Extensions/PromptTemplates.Prompty/Prompty/Types/ApiType.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Prompty.Core.Types
-{
-    public enum ApiType
-    {
-        Chat,
-        Completion,
-        Image,
-        Embedding
-    }
-}
