--- dotnet/src/Extensions/PromptTemplates.Prompty/Prompty/Types/ModelType.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Prompty.Core.Types
-{
-    public enum ModelType
-    {
-        azure
-    }
-}
