--- dotnet/src/Extensions/PromptTemplates.Prompty/Prompty/Types/ParserType.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Prompty.Core.Types
-{
-    public enum ParserType
-    {
-        Chat,
-        Embedding,
-        Completion,
-        Image
-    }
-}
