--- dotnet/src/Extensions/PromptTemplates.Prompty/Prompty/Types/RoleType.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Prompty.Core.Types
-{
-    public enum RoleType
-    {
-        assistant,
-        function,
-        system,
-        tool,
-        user
-    }
-}
